hell_mods for levels 60-65 made accurate
Note that this commit will result in some players who are levels 60 to 64 to end up one level higher than they were, as it changes the amount of experience for these levels.  Quarm shouldn't have a problem with this since we're still in classic era.  TAKP on the otherhand will have this problem, so I didn't push this there.  However these are the precise hell_mods and TAKP needs to have them eventually.