Pseudo-Classic Resists
This commit alters the resist spell method to try and mimic the classic, pre-September 2002 resists while keeping the 0-200 scale and -resist adjusts on spells.  Switching to the 0-99 scale would be too disruptive and troublesome to implement and would require a lot of database changes, so maintaining much of the PoP era system was desirable.  These changes are all gated to late Luclin and earlier.

Changes in this commit:

* Lifetaps on client targets are unresistable.
* Resist hardcap of 250 in Kunark, 350 after Kunark.
* Softcap at 200, overcap returns are half.
* Dragon breaths with a lure component hardcoded to be -100 instead of -150.
* Immolating Breath and Stun Breath hardcoded to be granted resist modifiers.
* Enchanter NPCs will be able to land spells on high resist targets.
* Partial hit damage is reduced for client casters. (it's kept at 50 and under levels)
* Partial damage taken by clients is increased.

I ommited the resist floors due to how disruptive those would be.  In some scenarios the resists are knowingly more leniant (higher caps) in this than in the Kunark decompile because the hardcap back then was rather harsh and the effectiveness per point of resist is reduced in many cases due to the 200 scale, requiring more resist to reach the same mitigation as classic.  I also believe that Sony likely altered the resist function at Velious launch to make resisting dragons easier and I can't say how exactly, but the obvious move would have been to raise the hardcap so that's what I did.