Fix fishing using the wrong heading
The incorrect heading was making it so your fishing line was being cast in the wrong direction.  For example standing East on the east freeport docks was casting the line to the southwest

Tested by both confirming multiple locations was able to fish vs not fish correctly.  There's a little funkiness on the EFP docks still because the line of sight check doesn't seem to check the dock. That's probably fine though. This at least fixes the majority of the issues.

Also double checked the headings in the logs, the locs it was trying to check, etc