Add support for any lighted item to shine from inventory slots
This will include the suggestion made for Staff of Null as well as other lighted items like candles, etc

This was working with a certain set of items like Greater Lightstone already, but I think it's hardcoded into the client to update those and not others.  I couldn't find a reason why the light was updating via server code for Greater Lightstone when it was put into an inventory slot using logs.

Suggestion thread: https://discord.com/channels/1133452007412334643/1277581221697814538