Pseudoclassic Resists Updated
This commit makes the classic era resists that I had previously implemented much more classic-like.  The previous commit was very incomplete and did not handle many Velious and Luclin boss spells and this is needed to fix that.  I've since better learned how resists were.  I will outline the changes because it probably looks strange as everybody is used to the new resist system and doesn't remember what it used to be like.

First it's important to understand that in classic, resists rolled a 0-99 instead of a 0-199.  Classic resists were more potent and in some ways resists became more unforgiving after the September 4 2002 revamp.  Post revamp, non-lure spells landed on players easier and also spells landing on NPCs became easier, which was a large DPS boost to caster players.  This meant that the all-or-nothing immunity threshold was 99 instead of 200.  So generally spellcaster NPCs in classic were actually much less threatening.  It was much easier to become fear immune to Nagafen for example and you did not need much resist gear.  Lured spells however were often harder to resist.

I could not apply most of this logic to NPCs because when Sony changed the roll to 200 they also added resists to many NPCs and applying this stuff to NPCs would require many database edits plus it's difficult to impossible to know which NPCs would need to be edited.  So players will still benefit from being able to land spells easier on most NPCs.  This commit should be not be considered a comprehensive reversion and it omits several things, most notably the lull and bard mez MR overrides (lull was near useless at higher levels in classic) and a collection of more trivial things.

Classic eras had resist caps.  Post revamp the cap is very high, 500.  In classic the PC resist cap -- against non-lures -- was 99.  Yes, 99.  I'm certain of this because the decompiles show it explicitly and the logs agree with this.  This would be the equivalent to 200 in the new system.  These caps were not apparent in the UI but they existed.  A result of this cap is that DD partial hit damage could not be mitigated beyond 50% average damage.

Classic also had resist floors.  It was not possible to debuff resists to 0 in classic.  I have implemented these floors (on NPCs and PCs) but since we're using a 200 roll I had to double them to make the resist rates the same as it was in classic.  I excluded tick saves when applying the floors even though they did apply to tick saves in classic because they would heavily nerf charm.  Any servers who want real accuracy will need to make the floors apply to tick saves.  Charm was much less powerful in classic.

In classic, lures did a separate resist roll, and this is why resists over 99 were still relevant back then.  The -resist adjust field in the spell data did not exist until September 4 2002 and before then spells had negative charisma in the first slot to denote lure behavior.  Lures back then would subtract an amount from the target's resist then halve the remainder and then apply a hardcap.  There were four types of lures back then: -3, -4, -5, -6.  For the most part resists against lure spells back then capped around 180-190 before level difference is factored in, so add another 50 to that against level 70 bosses.

I made CHA -6 resist spells resist at the same rate as they did in classic, with the same cap which is 45%.  Quarm players will notice that Kunark dragon AoEs will now resist less and this is intentional.  My previous logic was too lenient.  These became much easier to resist post-revamp because players can become immune after around 350-400 resist using bards in the new system.

-4 lure spells on the other hand had a high cap in classic and post-revamp Sony neglected to flag them not-partialable (all classic lures could not partial hit), so in PoP these were very hard to resist after previously being rather easy to.  The revamp made some encounters MUCH harder. (e.g. Lady Nev)  I made these spells no-partial but still somewhat harder (by 10%) to resist than in classic for a few reasons: first they became much harder to resist after the revamp as mentioned; second Sony arguably made these too easy to resist (the cap was 75%) in classic and in some cases the lure effect made the spells resist MORE than if they were just normal DDs (they still will); third players are still benefiting from some of the post-revamp changes.

When saving against non-lure spells I multiplied player resists by 1.5 because of the 200 roll.  Why not 2x?  Because resists were largely useless in classic and I didn't want to make them useless.  Also players still benefit from the lower NPC resists so I didn't want player resists buffed too much.

I added classic's partial damage formula but it requires more resist to reach 50% mitigation due to the 1.5x multiplier.

Players could not attain 100% spell immunity in classic and there is now a 1% minimum chance to land any spell on players.