Ceticious Cloud spell hardcodes
This spell is Severilous and Wuoshi's AoE.  In all of our eras (including PoP) this spell did no damage.  Before the September 4 2002 resists revamp the spell was lure and difficult to resist.  Sony may have hardcoded it to not do damage because the spell data seemingly had a mistake in it which is the spell has no recast delay on it, which means the dragons casted it every few seconds.  Why they didn't just put a delay on it to fix it instead I don't know.

I made the spell do 1 dmage instead of zero because if set to zero damage the server doesn't send a spell packet to the client due to the way our logic works it would require more "if spell_id ==" crap to make it send packets at zero damage and I didn't want to clutter up the code with more of it.