Instance boot 0 hp bug fix (#315)
* Check HP before Instance Boot

PQ currently has a bug in instances where sometimes you can hit a raid mob exactly as it dies and it can trigger being booted completely out of the instance. In some scenarios, this can be quite disruptive, and has become a very common issue for rangers with archery in particular.

This change adds inside AddToHateList  to skip the zone boot if the mob has 0 HP. In general, the game shouldn't be trying to add people to the hate list of a dead mob, but this check prevents the bug when it does happen.

I have never seen a pet trigger this bug, but added the check to the pet portion anyway for consistency.

* Remove the new check on pets

In testing, pets did not seem to cause their owner to get booted from the zone at all, either with OR without this change.

Since I could not properly test the pet conditional, I'm reverting that part of the HP check.