Fix VP Zone Out Ports (#320)
This fixes an issue with the zone out pads from VP -> Freeport, PoSky, etc.

These ports previously relied on the "double zone" bug to work because the first "if" check did not have code to handle same vs different destination zones.

This commit adds same vs different zone checks to that if branch and also does a minor refactor to remove bool is_same_zone strncmp as it is redundant with zone id == zone->GetZoneID()